Wallet.tmp/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pytest==8.3.3
httpx==0.27.2
cryptography==42.0.8
alembic==1.14.0
psycopg[binary]==3.2.9
SQLAlchemy==2.0.36
PyJWT[crypto]==2.9.0
